bpf: Decouple BPF_EVENTS from FTRACE and EVENT_TRACING
BPF_EVENTS controls inclusion of bpf_trace, which is also responsible for
tracepoint programs. In order to use tracepoint programs without needing to
enable FTRACE and EVENT_TRACING, make it possible to enable BPF_EVENTS
with just TRACEPOINTS as the only "event" type.

Even though this doesn't necessarily make it possible to use a tracepoint
program (since this doesn't require all tracepoints to be compiled), it is
enough to make the gpuWork eBPF program load successfully.

The gpuWork program is part of gpuservice in Android. When gpuWork fails to
load, gpuservice immediately kills itself on an assert. This occurs when
gpuservice starts up, causing gpuservice to crash continually in the
background.

The gpuWork program itself is not needed for any crucial functionality, and
neither is gpuservice itself to be honest. That being said, gpuservice is a
mainline module that is updated as part of Google Play system updates, and
when any mainline modules continually crash after a Google Play system
update, a reboot is triggered to rollback to the previous Google Play
system update:
  12-21 22:47:01.633  1655 11184 I ShutdownThread: Rebooting, reason: Rollback staged install

Hence, we need this so that the gpuWork eBPF program loads successfully.

Change-Id: Ia30822bdac0e67385efeea7e4e7e4c12a62764bd
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>